+++ packages/react-components/react-table/src/components/DataGridRow/useDataGridRow.tsx
@@ -23,7 +23,9 @@ export const useDataGridRow_unstable = (props: DataGridRowProps, ref: React.Ref<
   const columnDefs = useDataGridContext_unstable(ctx => ctx.columns);
   const selectable = useDataGridContext_unstable(ctx => ctx.selectableRows);
   const selected = useDataGridContext_unstable(ctx => ctx.selection.isRowSelected(rowId));
-  const tabbable = useDataGridContext_unstable(ctx => ctx.focusMode === 'row_unstable');
+  const focusMode = useDataGridContext_unstable(ctx => ctx.focusMode);
+  const compositeRowTabsterAttribute = useDataGridContext_unstable(ctx => ctx.compositeRowTabsterAttribute);
+  const tabbable = focusMode === 'row_unstable' || focusMode === 'composite';
   const appearance = useDataGridContext_unstable(ctx => {
     if (!isHeader && selectable && ctx.selection.isRowSelected(rowId)) {
       return ctx.selectionAppearance;
@@ -56,12 +58,13 @@ export const useDataGridRow_unstable = (props: DataGridRowProps, ref: React.Ref<
     {
       appearance,
       'aria-selected': selectable ? selected : undefined,
+      tabIndex: tabbable && !isHeader ? 0 : undefined,
+      ...(focusMode === 'composite' && !isHeader && compositeRowTabsterAttribute),
       ...props,
       onClick,
       onKeyDown,
       children: null,
       as: 'div',
-      tabIndex: tabbable && !isHeader ? 0 : undefined,
     },
     ref,
   );