+++ packages/react-components/react-toast/src/components/Toaster/useToaster.tsx
@@ -1,5 +1,11 @@
 import * as React from 'react';
-import { ExtractSlotProps, Slot, getNativeElementProps, resolveShorthand } from '@fluentui/react-utilities';
+import {
+  ExtractSlotProps,
+  Slot,
+  getNativeElementProps,
+  isHTMLElement,
+  resolveShorthand,
+} from '@fluentui/react-utilities';
 import { useFluent_unstable as useFluent } from '@fluentui/react-shared-contexts';
 import type { ToasterProps, ToasterState } from './Toaster.types';
 import { TOAST_POSITIONS, ToastPosition, useToaster } from '../../state';
@@ -14,18 +20,49 @@ import { ToastContainer } from '../ToastContainer';
 export const useToaster_unstable = (props: ToasterProps): ToasterState => {
   const { offset, announce: announceProp, ...rest } = props;
   const announceRef = React.useRef<Announce>(() => null);
-  const { toastsToRender, isToastVisible } = useToaster<HTMLDivElement>(rest);
+  const { toastsToRender, isToastVisible, pauseAllToasts, playAllToasts, tryRestoreFocus } =
+    useToaster<HTMLDivElement>(rest);
   const announce = React.useCallback<Announce>((message, options) => announceRef.current(message, options), []);
   const { dir } = useFluent();
 
   const rootProps = getNativeElementProps('div', rest);
 
+  // Adds native HTML focusin/focusout listeners
+  // https://github.com/facebook/react/issues/25194
+  const focusListenerRef = React.useCallback(
+    (el: HTMLDivElement | null) => {
+      if (el) {
+        el.addEventListener('focusin', e => {
+          if (
+            isHTMLElement(e.currentTarget) &&
+            !e.currentTarget.contains(isHTMLElement(e.relatedTarget) ? e.relatedTarget : null)
+          ) {
+            pauseAllToasts();
+          }
+        });
+
+        el.addEventListener('focusout', e => {
+          if (
+            isHTMLElement(e.currentTarget) &&
+            !e.currentTarget.contains(isHTMLElement(e.relatedTarget) ? e.relatedTarget : null)
+          ) {
+            playAllToasts();
+            tryRestoreFocus();
+          }
+        });
+      }
+    },
+    [playAllToasts, pauseAllToasts, tryRestoreFocus],
+  );
+
   const createPositionSlot = (toastPosition: ToastPosition) =>
     resolveShorthand(toastsToRender.has(toastPosition) ? rootProps : null, {
       defaultProps: {
+        ref: focusListenerRef,
         children: toastsToRender.get(toastPosition)?.map(toast => (
           <ToastContainer
             {...toast}
+            tryRestoreFocus={tryRestoreFocus}
             intent={toast.intent}
             announce={announce}
             key={toast.toastId}